+++ src/DeploySection.php
@@ -1,0 +1,9 @@
+<?php
+namespace TheRat\SymDep;
+
+class DeploySection
+{
+    const INSTALL = 'install';
+    const CONFIGURE = 'configure';
+    const LINK = 'link';
+}