+++ src/FileHelper.php
@@ -1,7 +1,13 @@
 <?php
 namespace TheRat\SymDep;
 
+use Deployer\Task\Context;
+use Deployer\Type\Result;
 use Symfony\Component\Process\Exception\ProcessFailedException;
+use function Deployer\isDebug;
+use function Deployer\run;
+use function Deployer\within;
+use function Deployer\writeln;
 
 /**
  * Class FileHelper
@@ -10,25 +16,40 @@ use Symfony\Component\Process\Exception\ProcessFailedException;
  */
 class FileHelper
 {
+    public static function copyFile($src, $dst)
+    {
+        $dstDir = dirname($dst);
+        $result = false;
+        if (!self::fileExists($dst)) {
+            if (!self::dirExists($dstDir)) {
+                run("mkdir -p \"$dstDir\"");
+            }
+            run("cp \"$src\" \"$dst\"");
+            $result = true;
+        }
+
+        return $result;
+    }
+
     /**
      * @param string $srcFilename
      * @param string $dstFilename
-     * @param string $mode
-     * @param array  $copyOnce
+     * @param string $mode     Example +x, 0644
+     * @param array  $copyOnce Regexp array
      * @return string
      */
     public static function generateFile($srcFilename, $dstFilename, $mode = null, array $copyOnce = [])
     {
-        $mode = !is_null($mode) ? (string) $mode : null;
+        $mode = !is_null($mode) ? (string)$mode : null;
 
         $copyOnce = array_map(
             function ($value) {
-                return env()->parse($value);
+                return Context::get()->getEnvironment()->parse($value);
             },
             $copyOnce
         );
 
-        if (in_array($srcFilename, $copyOnce) && fileExists($dstFilename)) {
+        if (in_array($srcFilename, $copyOnce) && self::fileExists($dstFilename)) {
             !isDebug() ?: writeln(sprintf('File "%s" skipped, because is in copyOnce list', $srcFilename));
 
             return '';
@@ -37,7 +58,7 @@ class FileHelper
         $dstDir = dirname($dstFilename);
         if (!self::fileExists($srcFilename)) {
             throw new \RuntimeException(
-                env()->parse(
+                Context::get()->getEnvironment()->parse(
                     'Src file "'.$srcFilename.'" does not exists'
                 )
             );
@@ -51,7 +72,7 @@ class FileHelper
         }
 
         $content = run(sprintf('cat "%s"', $srcFilename));
-        $content = env()->parse($content);
+        $content = Context::get()->getEnvironment()->parse($content);
         $command = <<<DOCHERE
 cat > "$dstFilename" <<'_EOF'
 $content
@@ -93,7 +114,7 @@ DOCHERE;
         foreach ($templateFiles as $src) {
             $name = str_replace($srcDir, '', $src);
             $dst = sprintf('%s%s', $dstDir, $name);
-            $res = self::generateFile($src, $dst, null);
+            $res = self::generateFile($src, $dst, null, $copyOnce);
             if ($res) {
                 $result[] = $dst;
             }
@@ -104,34 +125,55 @@ DOCHERE;
 
     /**
      * @param string $dir
-     * @return mixed
+     * @param string $workingPath
+     * @return bool
      */
-    public static function dirExists($dir)
+    public static function dirExists($dir, $workingPath = null)
     {
-        $cmd = sprintf('if [ -d "%s" ]; then echo true; fi', $dir);
-
-        return run($cmd)->toBool();
+        return self::runWithin("if [ -d \"$dir\" ]; then echo true; fi", $workingPath)->toBool();
     }
 
     /**
      * @param string $filename
-     * @return boolean
+     * @param string $workingPath
+     * @return bool
      */
-    public static function fileExists($filename)
+    public static function fileExists($filename, $workingPath = null)
     {
-        $cmd = sprintf('if [ -f "%s" ]; then echo true; fi', $filename);
-
-        return run($cmd)->toBool();
+        return self::runWithin("if [ -f \"$filename\" ]; then echo true; fi", $workingPath)->toBool();
     }
 
     /**
      * @param string $filename
-     * @return boolean
+     * @param string $workingPath
+     * @return bool
      */
-    public static function isWritable($filename)
+    public static function isWritable($filename, $workingPath)
     {
         $cmd = sprintf('if [ -w "%s" ]; then echo true; fi', $filename);
 
-        return run($cmd)->toBool();
+        return self::runWithin($cmd, $workingPath)->toBool();
+    }
+
+    /**
+     * @param string $command
+     * @param string $workingPath
+     * @return Result
+     */
+    public static function runWithin($command, $workingPath = null)
+    {
+        $result = null;
+        if (is_null($workingPath)) {
+            $result = run($command);
+        } else {
+            within(
+                $workingPath,
+                function () use ($command, &$result) {
+                    $result = run($command);
+                }
+            );
+        }
+
+        return $result;
     }
 }