+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 * CopyDbCommand - copy database command for test
 * documentation about **test** and **_secret.yml**
 
+[3.3.0]
+* symfony 3 as default recipe
+* file helper with closure command and specific dir
+* deploy on test like master
+* new build-box script
+* prepare list command
+
 [3.2.0]
 * functions as static methods
 * update user crontab, update nginx.conf in UpdateConfig class